movie-reco-api/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
gunicorn==23.0.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.0
psycopg2-binary==2.9.11
celery==5.4.0  
redis==5.0.8
requests==2.31.0
python-decouple==3.7
drf-yasg==1.21.11
pytest==9.0.1
pytest-django==4.11.1
faker==38.2.0
requests-mock==1.12.1